fix: use type field not kind when sending to homebase
In a previous PR merge, the field was changed from type to kind to follow
the spec we had on Notion. However, Homebase is not yet ready for this new
field and this resulted in a lot of "undefined" appearing in the Snyk Projects
page (when importing workloads) due to missing type. Reverting back to type
until Homebase is ready for the right parameter name.