fix: add egg deploy and cleanup files and scripts
The egg has until now been sitting in the default namespace.
This is considered bad practice, and none of our customers would do it like that.

Create our own namespace, where all of our resources will be stored.
This nicely isolates the monitor from the rest of the cluster resources.

Configure USER_LOCATOR from env var -- this way the monitor is able to read it and use it to associate workloads.

Also simplify deployment and rename resources.