fix: send only parent workloads
The monitor was initially written to send every workload that is a direct parent of a pod.
However, this is not the correct approach, and it results uploading loads of temporary objects (temporary in the sense that they are tied to a larger object, like a deployment, so a re-deployment effectively erases the object).

For example, if we had this structure:
Deployment -> 2x ReplicaSet -> Pods
We would be sending the 2x ReplicaSets. Instead, we want the Deployment!

This change attempts to recursively scan up until it finds the topmost workload -- and this is the one that gets sent to homebase.
The lookup stops whenever there are no more parents, or whenever it finds an unsupported workload.

For example:
Deployment -> UnsupportedWorkload -> ReplicaSet -> Pod
If we start from the Pod and look up, we will stop at the UnsupportedWorkload and consider the ReplicaSet the workload that can be inspected.