feat: add support for developing with Tilt
[Tilt](https://tilt.dev/) is a handy development tool for Kubernetes
projects.

What the current configuration supports when you run `tilt up` is:

* Building the Docker image locally
* Running the Helm chart, but swapping out the image reference for the
local one Tilt just build
* Presenting a handy browser and CLI UI which streams the log output
from the pods

More interestingly, Tilt will watch for any changes and reload the
updated parts of the application. For instance:

* If the Dockerfile or dependencies change it will rebuild the
image and update the running deployment
* If the application source code changes it will sync it to the running
container for even faster feedback
* If the Helm Chart configuration changes it will redeploy the updated
resources

The following relies on the Helm chart work, but I left in a separate
commit to make the proposed change clear. Happy to chat through, demo
how this works, and then discard or merge as we see fit.