fix: ensure workload props have values
According to the k8s docs, some properties of a workload have optional values.
The changes here ensure that we give default values to properties that may be missing.

We explicitly store the `kind` when reading a k8s object and we ensure optional values have at least empty defaults.