feat: watch on Pod changes
Remove initial scan from the monitor.
Add a Watch on the Pods of the cluster/namespace. This watch picks up all current Pods once the monitor starts (so it acts the same as a first-scan).
Ensures images (not tagged as latest) do not get processed more than once.

Also add watch permission to yaml files.
This allows the monitor to listen to changes to workloads (Pods).

Also pass namespace to the monitor.
The current namespace within the monitor cannot be inferred, so we have to pass it somehow.
This change prepares the monitor to receive it via an environment variable.