fix: add namespaced deployment
Currently the monitor can be deployed using kubectl and the YAML files in the whole cluster.
This may not work for every customer, as it requires permissions/access to the whole cluster.
This works well only if running in KinD or in Minikube, but folks will run into problems when they have access only to a single namespace.

Adding a namespaced deployment file and updated the README to reflect the change.
Some customers will be limited to a single namespace only, which means they can now deploy the monitor within that namespace!