feat: watch all workloads for DELETED events
When a workload is deleted, we need to notify Homebase so that it has an updated view of what can be imported.

The current approach of just listening for Pod changes doesn't work, because Kubernetes doesn't notify a Pod when it's parent owner is killed (e.g. when the Deployment is deleted).