fix: tone down image scans
Add a slightly more accurate way of deciding when to scan images in a Pod, reducing the number of extra scans by at least 1.

There is still one small caveat because of how Kubernetes fires events:
On deleting a Pod, Kubernetes first fires a MODIFIED event before firing a DELETED event.
We can't distinguish this MODIFIED event from a regular one so we don't know if a Pod is about to be deleted.
This results in an extra scan and then eventually deleting of the Pod. There is no fix for this (maybe we can do something smart about it in the future).