fix: upgrade k8s client to 0.10.2
This fixes a bug where a null reference happened due to the K8s API not setting an expected property.

The metadata-extractor.test.ts has been modified to reflect changes to metadata-extractor.ts, where a strict null check is now enforced in isPodAssociatedWithParent() so that the function does not throw.