test: if the setup fails, end the test process
today, if we fail to set up the test (tap.deployMonitor throws an error), the test case for deploying the snyk-monitor will fail but we'll continue to the next test cases.
if we fail setting up the test, clearly there's no reason to continue and we should make an effort at cleaning up the cluster we created, and regardless stop the test.

this might not be the cleanest way of achieving this, but tap's documentation is lacking and I could come up only with this.