feat: bail out early when there's no work to do on workload processing
when we process a workload, we get a unique list of images belonging to it.
we first pull these images.
the images we managed to pull successfully will be scanned.
the images we managed to scan successfully will have their results sent.

this commit introduces checks between these steps, so that if the count of images that were successful is 0, we will stop the processing, as there's nothing to do (no point to scan 0 images, then send their "results")