fix: clearer image metadata building
asserting buildImageMetadata will always be called with a populated (or at least defined) list of containerStatuses.
bailing out if this list does not exist, as we can't construct imageId data without it.
removing old technical debt accumulated to allow this function to be called from a delete image context, as it no longer participates in this flow.