chore: test the image we build using semantic versioning
stop building the image in our integration tests' setup.
instead, expect the image to be already built and available, and receive its tag in an environment variable.

in local or pull-request runs, build a temporary image with the "local" tag, and use it for the tests. (called from "npm pretest" rule)

in merge-to-staging scenario, build an image, tag it as snyk/k8s-monitor:<new-semrel-version>-candidate, and test it.
if everything passes, re-tag it as snyk/k8s-monitor:<new-semrel-version>-approved.

in merge-to-master scenario, fetch the job commit's latest tag and re-tag it to just snyk/k8s-monitor:<new-semrel-version>.