fix: log workload name instead of random log ID
Initially the idea of logging a random ID was to distinguish different workload processing flows as many of them ran concurrently and created log messages.
Sometimes it was difficult to follow for example where 3 of the same "Scanning pulled images" messages come from.

Logging a random ID actually doesn't give us any meaningful information; instead we can log the workload name, which is already unique as we opt to scan an image only once.
It should be much clearer to also figure out which workloads have been processed by the monitor and we can verify that they must appear in Snyk import UI.