fix: attach PersistentVolumeClaim to the monitor in Helm/YAML
This is in preparation of using skopeo for pulling and saving images, so that they can be statically scanned later.
The PersistentVolumeClaim (PVC) is a request to Kubernetes to give it a disk volume of a certain size (and storage class, but we use only defaults).

The request is for 20Gi disk mounted at /snyk-monitor inside the container.