chore: unset PR-related env variables on "merge to staging" flow
CIRCLE_PULL_REQUEST, CIRCLE_PULL_REQUESTS, CI_PULL_REQUEST, CI_PULL_REQUESTS
if these env variables are set, semantic release will mistaken the run as a pull request and refuse to release.
they may be set because of how CircleCI interacts with PRs - it doesn't respond to PRs opening and trigger builds, instead, it starts running when pushes are made to any branch, and adds the PR variables if it detects a PR exists.