chore: attach host storage and use flags for static analysis
Create a 20Gi emptyDir to be used by the monitor.
The space will be used to temporarily store pulled container images, which are later to be statically analysed.

Deciding on host storage instead of managed attached storage to allow for reduced installation friction.

Also, using an environment variable passed with the deployment, hint to the monitor that static analysis may be used.
This would later allow us to experimentally enable static analysis (which we may consider from now on a beta feature).