feat: use skopeo and static analysis when flag is set
Treat static analysis as a feature that can be enabled with a flag passed as an environment variable.
When static analysis is enabled:
- replace docker pull with skopeo copy
- pass static scanning options to the plugin to hint scanning the image statically
- delete the pulled image if it was previously pulled statically

These tasks are as non-invasive as possible and do not try to branch to two paths (static vs dynamic).

To use skopeo, we are now using multiple stages in our Dockerfile:
The first stage builds skopeo from source, using a golang+alpine base image.
The second stage is as it is currently, but only copies the skopeo binary and additional files that it needs.