chore: reorder scanning logs
Right now if we don't scan images we get logs like this (in this order):
- successfully scanned 0 images
- no images were scanned, halting scanner process

Reoder the logs so that when 0 images are scanned, we don't show the success message but say only "halting the scanner process".
When 1 or more images are scanned, that's when we will show the "successfully scanned N images" log.