chore: split monitor setup for efficient deployment
The "deployMonitor" call right now both deploys the monitor and a few sample fixtures (deployment, replicaset, etc.).
For stress tests, we use the same deployMonitor() to load an instance of the monitor but we also initialise these extra fixtures which we don't check.

This change extracts the fixture creation in a separate function, so that deploying the monitor is its own contained step.
Also reduced some repetitive try-catch clauses and moved them inside deployMonitor().

Also, extracting out Homebase checks to its own file because this functionality can (and will) be used by other tests.