test: add static analysis stress tests
Add new build jobs that run as part of opening a PR.
The jobs run "stress" tests using a variety of images for currently supported package managers to ensure that the monitor can scan the same workloads between releases (catching regressions in scanning).

Also, extract repetitive blocks in CircleCI config so that it's easier to read and maintain.