fix: pull images to uniquely-named files
Currently there is a bug that happens rarely but once it does it crashes the Node.js process.
It occurs when two workloads which contain an identical image are pulled at the same time.
The images are pulled to the same filesystem name and the two processes react poorly with each other, resulting in a crash.

This fix adds a unique identifier to every file system name so that clashes do not occur.