feat: set readOnlyRootFilesystem and change temp storage mount path
Enable readOnlyRootFilesystem on the deployment files but also change the temporary storage mount path.
Regardless of where we store pulled images, skopeo seems to pull temporary files under /var/tmp and this behaviour is not configurable.
Because of that we cannot enable readOnlyRootFilesystem unless we also mount our temporary storage there.