feat: set readOnlyRootFilesystem, change temp storage mount path
Upgrade and pin to snyk-docker-plugin version 1.34.0.

The static scanning options now accept a file path to a temporary directory for saving snyk-docker-plugin files.
To be used in combination with readOnlyRootFilesystem and the directory for pulling temporary images.

Enable readOnlyRootFilesystem on the deployment files but also change the temporary storage mount path.
Regardless of where we store pulled images, skopeo seems to pull temporary files under /var/tmp and this behaviour is not configurable.
Because of that we cannot enable readOnlyRootFilesystem unless we also mount our temporary storage there.