feat: ignore workloads that were deleted during their scanning
Due to the async nature of processing workload events, it's possible to still send a scan result of a workload that got deleted while in the process of being scanned, thus appearing as something importable in Snyk UI.

This change attempts to track workload that are being scanned and workloads that got deleted while in the process of scanning. This tracking helps us decide whether or not to send the scan result: we would skip sending if the workload got deleted in the middle of the scan.