feat: use dumb-init to wrap the main process
PID 1 in Linux has special treatment, and won't react as most processes do to certain signals (kill, int, term...).
this means when the Kubernetes-Monitor gets an interrupt from Kubernetes, it probably will ignore it and later on be killed forcefully without a graceful attempt at shutting down.

this is probably insignificant for now, but may be more significant as time goes by when we split to worker/manager pattern and maybe attempt some caching or syncing between adding/deleting workloads.

detailed article here https://engineeringblog.yelp.com/2016/01/dumb-init-an-init-for-docker.html