fix: `await` on workload metadata POSTs
we see a lot of the workload metadata requests getting hung up on their way to the upstream service.
this appears to happen more frequently when the Kubernetes-Monitor starts, especially on more crowded clusters.

after some investigations, it looks like we're sending a lot of this kind of requests and possibly more than the event loop can handle with ease.
this is made worse by not `await`ing on these requests, handling more things before sending them, causing them to time out (that's half a guess by the way).

testing the addition of `await` on our dev cluster appears to help.