chore: allow using things "before" defining them in tests code
this eslint rule is either incorrectly described, or bugged in its behaviour.
at the moment it blocks things such as:

function f1(): {
  f2()
}

function f2(): {}

claiming that f2 is used before it is defined.
this is incorrect because while f2 is referenced in an earlier line, it is invoked only when f1 is invoked, not when f1 is defined.