fix: build image only once during tests
Adjust the CircleCI stages so that the container image is built only once and reused for all the tests (per workflow). Allows us to test the same image under a variety of scenarios (integration tests, package manager tests, -anything new-).

Additionally, include the commit hash when notifying Slack that a job failed as it is easier to track where a problem comes from.

Also, split the test job to run unit tests and integration tests in parallel, to allow for quicker builds.