chore: always notify Slack on failure
Use CircleCI's "when: on_fail" as part of a command to run in order to always notify Slack on job failures.
The problem up until now was that long-running commands may time out and CircleCI just interrupts them so we never reach the part of the script where it should notify Slack.