test: discrepancies between owner spec and child statuses
metadataExtractor.buildImageMetadata accepts a workload's metadata and a list of container statuses, belonging to containers in a pod running said images.
this function attempts to build image metadata based on the owner's spec (for example a deployment), which contains the data about "what should be", as well as the container statuses for that pod, which contain data about "what actually is".

some discrepancies may occur between "what should be" and "what actually is".
an example we've stumbled upon happens when sidecar containers are injected (through an admission controller), causing the deployment's spec to contain a single container, but the statuses to include the injected containers.

this results in an error in buildImageMetadata that relies every container that appears in the statuses list to also appear in the spec.

this test proves the bug exists.