chore: instruct containers to sleep when creating a deployment from an image
right now the containers will use their default command, usually resulting in nothing or a loop of crashing.
this creates race conditions in the Kubernetes Monitor and adds some sporadic behaviours to our tests.
for the consistent tests, let's make the behaviour of the containers consistent - just instructing them to sleep.