chore: eks/ecr integration testing preparations continued
- decoupling cluster creation from making the image accessible to the cluster
EKS clusters, even if existing, will need access to the image being tested.
this is most easily achieved by pushing the image to ECR, just like in KinD we load the image to KinD.

- loadImageInCluster may decide on an updated image name for testing
this is because when we push the image to the remote cluster, its name/tag may change to include the registry's prefix.

- clean an environment instead of deleting it
for clusters we don't want to delete (EKS), just clean up.

- more generic decoupling from KinD

- adding the EKS test to CircleCI config