chore: bump KinD version
as part of the version bump, KinD is deprecating the "get kubeconfig-path".
instead, they alter supposedly the default kubeconfig file, and recommend to use it through the context flag.
this makes our testing infra a bit more complex because we're using sometimes the Kubernetes API client and sometimes kubectl.

an easy workaround I have chosen was to dump kubeconfig's content to a new file, and just export its location as we used to.