fix: retry and throttle requests to k8s-api
When building a workload's metadata, we send requests to the Kubernetes-API to get additional details (a workload's parent).
Some of these requests fail with an HTTP response code of 429 ("too many requests") which is basically the Kubernetes-API telling us its busy at the moment.
These responses also may contain a "Retry-After" header, that according to the HTTP standard is either a non-negative decimal integer or a date, serving as an instruction by the server for when the client should try next.
https://tools.ietf.org/html/rfc7231#section-7.1.3

This commit adds a retry mechanism to our Kubernetes-API requests in case we get such a 429 response.