chore: new helper function to wait for the default service account
when a new KinD cluster is created (this possibly applies to all Kubernetes clusterss), it takes a while for the default service account entity to be created.
during this time, deploying workloads that require it won't work.
this function allows its called to wait indefinitely until the serviceaccount is created.