test: system test
introducing a new system test infrastructure to the Kubernetes Monitor.

this kind of tests focus on the Kubernetes-Monitor's code, and its interactions with Kubernetes, without letting it run completely in an end-to-end fashion.

this means that:
1. we're losing the Monitor's "natural" runtime environment (inside the cluster)
2. we're gaining the ability to test for implementation details we otherwise couldn'tsuch as the content of the payload the Kubernetes Monitor sends out (as opposed to testing its affect on an external upstream)

the new test creates a KinD cluster, but runs the Kubernetes-Monitor _locally_ (not from inside the cluster). it uses Nock to assert outgoing requests.
right now it only uses a java deployment, because it's distroless, saving us the trouble of forcing an OS on Skopeo's pull (which would try to get an Ubuntu image for MacOS... and fail.)