fix: assert file system path for skopeo copy is truthy
pullImages implicitly relies on the list of pullable images it receives to have fileSystemPaths for each image, as a valid path in the filesystem where we can instruct Skopeo to pull images to.
right now, it accepts undefined values as these paths, resulting in calls to Skopeo creating a file named "undefined" on the filesystem.
this commit amends this behaviour by asserting the fileSystemPath is truthy before calling Skopeo.