fix: exit the process on unhandled rejections
unhandled rejections are usually an indication of something going pretty wrong with the code, with no way to recover.

so far, we've been silently ignoring those (and only logging) because we didn't know when these things would tend to happen on the Kubernetes Monitor.

today we're a bit more confident about the unhandled rejections we're seeing, and we believe it's better to exit the process and let the pod restart.

also removing the "promise" prints as they always end up as an empty object, and never provide any interesting data.