chore: relax checks on runAsUser and runAsGroup
To allow our tests to check some security properties at runtime, we need to relax specifically runAsUser and runAsGroup.
For OpenShift the UID/GID is a very high number but it is also randomised, so there is no need for us to try and predict it and be exact in our tests.
The change will allow any sufficiently high UID/GID (>=10001), which should still be a good measure of following security best practices.