fix: retry skopeo copy command
DockerHub being instable recently introduced new flakiness to our tests, which points towards our "skopeo copy" command being a bit too fragile.
this commit introduces a retry mechnanism on that command, always retrying to pull the image.

a more efficient solution would be inspecting the error causing the instability and retrying only on that, but this should be good enough as we'll waste only 2 seconds at worst.