chore: add OpenShift 4 integration tests
The tests require the oc CLI tool to authenticate to the cluster so it attempts to download it as part of running the tests.
To run the tests it requires setting up several environment variables that point to the cluster.

Also temporarily patch the YAML deployment to remove runAsUser and runAsGroup on deploying the snyk-monitor - those will be addressed in the future as part of Operator tests.