fix: explicitly set HOME env var and pass it to subprocesses
OpenShift deployments override HOME in the running Pod to point to "/" instead of the user's home directory "/srv/app".
This breaks skopeo from being able to pull images as it relies on having access to HOME.
This fix hardcodes/overrides HOME so that it allows the snyk-monitor to run on OpenShift.