chore: run OpenShift 4 integration tests in CI
Run the tests only on merging to staging, similarly to EKS tests and for the same reasons: requiring exclusivity of a single cluster, but also having the need to run the tests only when the product is going to be built with a new release.

The CI job also overrides /etc/hosts with an entry to the OpenShift cluster -- this is because we use a CodeReady Containers cluster which is usually meant for local development/testing, so some of its configuration is hardcoded.
On attempting to authenticate to the CRC cluster it returns an OAuth endpoints to use -- the endpoint is hardcoded to "api.crc.testing" hence it is not reachable from the outside world. The /etc/hosts override allows us to point this entry to the static IP address of our CRC cluster.