chore: introduce intermediary Operator files
The operator-sdk is used with its Helm plugin to generate these intermediary files, which would later be used to generate the final product (ClusterServiceVersion) for publishing for customers to use.
The intermediary files define what resources the Operator is controlling, and what permissions the Operator has. By default the operator-sdk generates a full permission set for us.

Alongside the intermediary files is a Dockerfile that can be used by the operator-sdk to build the Operator image. The image itself uses a special Helm Operator base image that handles the creation of our resources (the snyk-monitor Helm chart) for us.
The image is also used for publishing the Operator to customers and must be publicly available once built.

We are checking in the files to the repository to allow for fine-grained access and control of the final Operator. Specifically, the Community Operator Hub strongly encourages revisiting the permissions on the Operator and removing ones that are not needed for its functioning (e.g. accessing Secrets).
Since the SDK does not give us an easy way to control the RBAC permissions, checking in the files allows us to later trim down the unnecessary permissions.
Another point for doing this is the current unpredictability of the SDK, which produces completely different RBAC sets on Mac and on Linux.