chore: introduce intermediary Operator files
The operator-sdk is used with its Helm plugin to generate intermediary files, which would later be used to generate the final product (ClusterServiceVersion) for publishing to OperatorHub for customers to use.
The intermediary files define what resources the Operator is controlling, and what permissions the Operator has. By default the operator-sdk generates a full permission set for us. These intermediary files are not going to be checked in to the repository or generated dynamically because the final product (ClusterServiceVersion) contains everything we need. The ClusterServiceVersion will be introduced in a later change and will be hand-curated based on the intermediary files.

Alongside the intermediary files is a Dockerfile that can be used by the operator-sdk to build the Operator image. The image itself uses a special Helm Operator base image that handles the creation of our resources (the snyk-monitor Helm chart) for us.
The image is also used for publishing the Operator to customers and must be publicly available once built.
The Dockerfile is checked in because it is essential to build an Operator image every time we release a new version. This is because we want to embed default values in the Helm chart (e.g. pointing to the correct version of the snyk-monitor) and to align the Helm chart version and snyk-monitor version with the Operator version. This is the only supported way as all the versions increment in lockstep.