chore: add a ClusterRole to the Operator
The way the Operator works is in several ways: it can be deployed to a single namespace, to multiple namespaces, or to all (the whole cluster).
This is where there is a slight discrepancy in how we use the snyk-monitor: we deploy it to a single namespace yet we want access to scan the whole cluster.

The Operator SDK is not ready to understand this concept -- if you deploy an Operator to a single namespace then it assumes you need only a Role with permissions on that namespace.
However, on deploying our Operator, by default the snyk-monitor is installed with scope: Cluster, which means the Helm Operator needs to create a ClusterRole and ClusterRoleBinding for us. At this point the Operator will fail to do it.

This change is adding the necessary permissions to allow for the normal functioning of the snyk-monitor.