chore: integrate Operator creation in CI/CD
On every PR and merge to staging, run a build job to create the Operator intermediary files.
The script used to create these files also modifies some placeholder tags to ensure we can deploy the correct version of the Operator and snyk-monitor.

This change intentionally only creates the files -- they will be consumed by a later build job step -- generating the ClusterServiceVersion for publishing the Operator.