fix: hardcode HOME in the deployment of snyk-monitor
This ensures that HOME is set/overridden as an environment variable in the snyk-monitor container in both plain YAMLs and Helm.
This fixes an issue when deploying on OpenShift where the HOME env var points to / instead of /srv/app, even though we run the container as the user "snyk" whose home directory is /srv/app.
The wrong HOME (pointing at /) makes skopeo (the tool we use to pull images currently) look for image pull secrets (dockercfg) in the wrong place. By default skopeo looks under $HOME/.docker/config.json but when HOME points to the wrong place, this means it stops being able to pull from private registries. By hardcoding HOME in the deployment files, we are able to fix this problem.