fix: in case we fail to find a workload's parent, bail out
when looking for a workload's parent, if the workload is associated with a parent, but we fail to find it (meaning it was probably deleted), return "undefined" instead of the toppest-tiered parent we found so far.
this means buildMetadataForWorkload() will also return "undefined" and the podWatchHandler() on pod.ts will log a warning and stop acting, instead of using the middle-tier parent in a best effort.