fix: retry transmissions on certain network failures
when failing to transmit data to the Kubernetes-Upstream service, if the failure is a result of specific network errors, retry a certain amount of times.
otherwise, fail the request.

this is done because networks are often unpredictable and we see some socket hangups or unexpected network socket disconnections. in these scenarios we want to retry sending the request, so the important data isn't lost.