chore: remove package manager tests
These tests had the purpose of providing us with confidence that we can scan all supported package managers on a variety of images.
They are not bringing us much value, because we already test all package managers with the rest of our fixtures in the general integration tests.
The snyk-docker-plugin that we consume also has sufficient test coverage, so the package manager tests no longer serve their purpose.