chore: remove CREATE_CLUSTER environment variable in tests
Make the platform setup entries generic enough so that they do not throw and do nothing whenever an action is not needed.
This allows us to execute actions in order and not be worried that they will throw, and allows us to remove the CREATE_CLUSTER environment variable. The platform itself will decide whether it needs to create a cluster, whether to delete or clean it, etc.