chore: extract test tools into their own module
This allows to consolidate all the tools we use during tests into one place and make it easier to install them and clean them up.
This current change starts with skopeo and simplifies the system test, the idea is to do the same for all tools we use (kubectl, kind).