chore: delete test Operator from Quay on finishing OpenShift tests
The versioning scheme for our test Operators is 0.0.1-<commit sha>.
Quay orders releases alphabetically when deciding what is the latest Operator release, meaning that the commit SHA with the highest string value wins - in the current case it is a several weeks old release starting with the SHA f6162b...
This means we have been testing the wrong release with every OpenShift test run.

The Operator catalog does not give us an easy way to pin to a specific version of the Operator, it will always choose the latest.
Hence as a workaround we can delete the Operator from Quay once we are done testing it.