fix: simplify scanner module by removing the usage of class
The class was initially added to conceptually separate the scanning/processing logic from the rest of the code. This separation is in order to later transfer this piece of logic into a separate worker mechanism, e.g. a Kubernetes Job. This separation however never happened, and this is the only place in the code that uses a class but without an apparent reason.

Convert the class functions into plain async functions and export them in the scanner module.