fix: remove runAsUser/Group to support OpenShift 3
These values are automatically generated by OpenShift and cannot be the ones we have been enforcing until now.
Also update the Dockerfile to run with the numeric user and group -- this allows the deployment to proceed, otherwise Kubernetes cannot infer the UID of the user "snyk" and fails to deploy.