feat: use a library to scan RPM packages instead of an external tool
Upgrade the snyk-docker-plugin dependency which carries @snyk/rpm-parser.
This allows us to scan RPM packages directly instead of shelling out to a tool.

We no longer need to pass a path to a temporary directory for storing extracted files.
Also, we no longer need to build and embed the tool in the Dockerfile.