chore: validate both dep graphs are present in binaries tests
There is a possible race condition that can cause test flakiness when testing the binaries deployment with two dep graphs.

The reason is as follows:
A workload is available from Homebase when at least 1 dependency graph and its workload metadata are stored. Currently we don't do a check that asserts that both dependency graphs of a deployment are stored and the tests may fail.

This fix ensures that both graphs are stored before proceeding with the tests.