fix: fix permissions for registries.conf file in OpenShift 3
The .config directory in the product is shared between several entities:
- snyk protect creates it as part of storing metadata
- we mount registries.conf to override some image pulling properties

On OpenShift the mounting over an existing directory acts weird and makes the .config directory completely inaccessible to our user and breaking the registries.conf support that we have.
This fix pre-creates the directory with the right permissions in the Dockerfile so the OpenShift 3 behaviour is fixed.