chore: delete all Operators from Quay on clean up
Once Operator tests complete we currently have a script that tries to delete the last Operator from Quay.
However, in case of errors some Operators might have been stuck in the repository.

This change tries to delete all Operators from the Quay repo in an attempt to leave a clean environment.