feat: PodSpec is now read from Pods instead of the workloads
This is a workaround for cases where the workload is a Custom Resource, which snyk-monitor today does not support and cannot read them.
There are situations where we will misreport the securityContext because we tried to read it from the wrong workload.

The Pod's spec is an identical copy to the one defined in the workload, so we can use it for collecting the workload information that we need - this way we can track workload security configuration even if the Pods are managed by a Custom Resource.