docs: Update helm installation command
If any older version of the snyk helm repo is already present on the local machine, the current `helm repo add` command will not modify it, meaning the subsequent call to `helm upgrade --install` will use the latest _local_ chart, which may not be the latest _actual_ release. Adding `--force-update` as per this PR will ensure that the local repo is updated to include the latest actual release.