feat: add new templates for certifying the Snyk Operator
The certification is for the Red Hat Marketplace, as described in their documentation.

The files are currently stored in a separate directory to avoid conflicts with the existing snyk-operator directory (which is used by tests and CI/CD).

Some notable changes from the previous/existing Operator:
- Fixing permissions - the ClusterRole and ClusterRoleBindings are no longer applied via the Helm chart but instead are listed in the Operator metadata (the ClusterServiceVersion)
- New breaking change base image for the Operator (based on the Helm template Operator)
- Using RBAC proxy provided by Red Hat
- Different name for the Operator - snyk-operator-certified

There are still several TODOs not covered by this change:
- moving the certified images over to Quay
- tests should use and test the certified Operator
- CI/CD should use the certified Operator - build, test, push the image for publishing