fix: create PVC only if using the default PVC name
We should create a PVC only if the default PVC name has not changed.
If the name has changed, it means our users want to use their own pre-provisioned PVC.
In such cases it would be an error to create a PVC as it would result in a conflict - the PVC already exists.