chore: delete Operators from Quay after tests complete
Right now both the Operator on OpenShift and Operator on OLM tests depend on the same Operator being pushed to our Quay repository. Unfortunately the OpenShift tests delete the Operators from Quay once the test finishes, which interferes with the other test.

This ensures that the Operators are deleted only after both tests that depend on it complete.