chore: pass process.env when calling child_process.exec()
It seems like there is an issue with Jest where changes to process.env are not persisted through calls to child_process.exec(). This causes env vars like KUBECONFIG to be lost and to fail our tests. For now use a wrapper to explicitly pass process.env as a workaround.